ServerSentEvents/field-name-data.html

<!DOCTYPE html>
<html>
<head>
		<title>The steps to process the field given a field name and a field value depend on the field name: The field name is "data":</title>
  <meta charset="utf-8" />
  	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="/resources/WebIDLParser.js"></script>
	<script src="/resources/idlharness.js"></script>
</head>
<body>
	<h1>The steps to process the field given a field name and a field value depend on the field name: The field name is "data":</h1>
	<div id="log"></div>
  <script>
  	setup({timeout:1500});
    var s = new EventSource('http://html5.mosquito-fp7.eu/messages/t0075');
    vt = null;
    i = null;
    t1 = async_test("field name is 'data' and append any U+000A LINE FEED (LF) character to the data buffer: Append the next field value and data buffer on the same line.");
    t2 = async_test("field name is 'data' and append a single U+000A LINE FEED (LF) character to the data buffer: Append the field value to the data buffer.");
    t3 = async_test("field name is 'data' and append two U+000A LINE FEED (LF) characters to the data buffer: Send a new data buffer, append the field value to the data buffer.");

    s.onopen = function(e){ 
    	/* Connection opened. */
    };
    	
    s.onmessage = function(e){
    	var dat = e.data.split("\n");
    	console.log(e.data.split("\n"));
    	t1.step(function(){assert_equals(dat[1], "data bufferdata: no LF")});
    	t2.step(function(){assert_equals(dat[2], "one LF")});
    	t1.done();
    	t2.done();
    	if(i == null){
    		i=1;
    	}else{
    		t3.step(function(){assert_equals(dat[0], "two LF")});
    		t3.done()
    	};
    };
    
    //	console.log(s.testevent);
    s.onerror = function(e){
    	/* Connection closed. */
    };
    

  </script>
  </div>
</body>
</html>
